requirements.txt
pymongo
customtkinter
Pillow
python-dotenv
tkcalendar